+++ frontend/app/actions/Fallbacks.scala
@@ -1,48 +1,11 @@
 package actions
 
-import com.gu.memsub.subsv2.CatalogPlan.PaidMember
-import com.gu.memsub.subsv2.MonthYearPlans
+import abtests.CheckoutFlowVariant
 import com.gu.salesforce.PaidTier
+import configuration.Config
 import play.api.mvc.Results._
 import play.api.mvc.{Call, Cookie, RequestHeader}
 import play.twirl.api.Html
-import configuration.Config
-import views.html.joiner.form.paymentB
-import views.support.{CountryWithCurrency, IdentityUser, PageInfo}
-
-import scala.util.Random
-
-
-sealed trait CheckoutFlowVariant {
-  val testId: String
-  val identitySkin: String
-}
-
-object CheckoutFlowVariant {
-  val cookieName = "ab-checkout-flow"
-
-  val all = Seq[CheckoutFlowVariant](A,B)
-
-  def deriveFlowVariant(implicit request: RequestHeader): CheckoutFlowVariant =
-    getFlowVariantFromRequestCookie(request).getOrElse(CheckoutFlowVariant.all(Random.nextInt(CheckoutFlowVariant.all.size)))
-
-  def getFlowVariantFromRequestCookie(request: RequestHeader): Option[CheckoutFlowVariant] = for {
-    cookieValue <- request.cookies.get(cookieName)
-    variant <- CheckoutFlowVariant.lookup(cookieValue.value)
-  } yield variant
-
-  case object A extends CheckoutFlowVariant {
-    override val testId: String = "test-A"
-    override val identitySkin: String = "members"
-  }
-
-  case object B extends CheckoutFlowVariant {
-    override val testId: String = "test-B"
-    override val identitySkin: String = "membersB"
-  }
-
-  def lookup(name: String): Option[CheckoutFlowVariant] = all.find(_.testId == name)
-}
 
 object Fallbacks {
